ZeldaCharactersApi/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
pymongo==4.6.1
python-dotenv==1.0.0
